thermal-hal: Return empty vector when a sensor is not supported
Right now the thermal HAL will return a temperature, temperature
threshold and cooling device entry with a NAN value for a sensor that is
not supported. This will break the VTS test case, which expects an empty
vector in that case.

Change the thermal HAL to return empty vector in case the sensor is not
supported.

Also populate the thresholds value for CPU, GPU, Battery and skin
sensors for the V1.0 getTemperature API. Make sure that the threshold
values are also read in degree celsius.

Change-Id: I82ab1e9dd9672f1b93760e93b55e8943fc5acce0