thermal-hal: Notify clients when they read temperature
Certain sensors like thermistor's, can read temperature higher than the
threshold even before a threshold violation uevent. This is because
certain thermistor's use comparator's to trigger threshold violation
independent of the ADC used to read the temperature. In those corner
case, thermal HAL can update the severity when a client is reading a
temperature and not notify them. So when an actual uevent occurs, client
notification will be skipped, because the severity wouldn't have
changed.

Update HAL to notify clients when the read temperature API see's the
threshold violation first before the uevent.

Change-Id: I30248b1364909a74b362fa732041c3f96917952c